+++ src/main/java/com/uzem/book_cycle/admin/dto/sales/SalesRequestDTO.java
@@ -5,7 +5,6 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
-import static com.uzem.book_cycle.admin.type.SalesStatus.AVAILABLE;
 
 @Getter
 @AllArgsConstructor
@@ -24,7 +23,7 @@ public class SalesRequestDTO {
 
     @NotNull
     @Min(value = 0, message = "가격은 0 이상이어야 합니다.")
-    private int price;
+    private Long price;
     @NotNull
     private BookQuality bookQuality;
 }
